move liquidate computations back to keeper package
